Cleanup modules structure and architecture. Split up types/struct creation from functions declaration and usage.
