+++ src/ldfd/api/distributions/distributions.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/bitswalk/ldf/src/common/logs"
 	"github.com/bitswalk/ldf/src/ldfd/api/common"
-	"github.com/bitswalk/ldf/src/ldfd/auth"
 	"github.com/bitswalk/ldf/src/ldfd/db"
 	"github.com/gin-gonic/gin"
 )
@@ -18,80 +17,12 @@ func SetLogger(l *logs.Logger) {
 	log = l
 }
 
-// Handler handles distribution-related HTTP requests
-type Handler struct {
-	distRepo   *db.DistributionRepository
-	jwtService *auth.JWTService
-}
-
-// Config contains configuration options for the Handler
-type Config struct {
-	DistRepo   *db.DistributionRepository
-	JWTService *auth.JWTService
-}
-
 // NewHandler creates a new distributions handler
 func NewHandler(cfg Config) *Handler {
 	return &Handler{
 		distRepo:   cfg.DistRepo,
 		jwtService: cfg.JWTService,
 	}
-}
-
-// CreateDistributionRequest represents the request to create a distribution
-type CreateDistributionRequest struct {
-	Name       string                 `json:"name" binding:"required" example:"ubuntu-22.04"`
-	Version    string                 `json:"version" example:"22.04.3"`
-	Visibility string                 `json:"visibility" example:"private"`
-	Config     *db.DistributionConfig `json:"config"`
-	SourceURL  string                 `json:"source_url" example:"https://releases.ubuntu.com/22.04/ubuntu-22.04.3-live-server-amd64.iso"`
-	Checksum   string                 `json:"checksum" example:"sha256:a4acfda10b18da50e2ec50ccaf860d7f20b389df8765611142305c0e911d16fd"`
-}
-
-// UpdateDistributionRequest represents the request to update a distribution
-type UpdateDistributionRequest struct {
-	Name       string                 `json:"name" example:"ubuntu-22.04"`
-	Version    string                 `json:"version" example:"22.04.3"`
-	Status     string                 `json:"status" example:"ready"`
-	Visibility string                 `json:"visibility" example:"public"`
-	SourceURL  string                 `json:"source_url" example:"https://releases.ubuntu.com/22.04/ubuntu-22.04.3-live-server-amd64.iso"`
-	Checksum   string                 `json:"checksum" example:"sha256:a4acfda10b18da50e2ec50ccaf860d7f20b389df8765611142305c0e911d16fd"`
-	SizeBytes  int64                  `json:"size_bytes" example:"2048576000"`
-	Config     *db.DistributionConfig `json:"config,omitempty"`
-}
-
-// DistributionListResponse represents a list of distributions
-type DistributionListResponse struct {
-	Count         int               `json:"count" example:"5"`
-	Distributions []db.Distribution `json:"distributions"`
-}
-
-// DistributionStatsResponse represents distribution statistics
-type DistributionStatsResponse struct {
-	Total int64            `json:"total" example:"10"`
-	Stats map[string]int64 `json:"stats"`
-}
-
-// getTokenClaims extracts and validates JWT claims from the request
-func (h *Handler) getTokenClaims(c *gin.Context) *auth.TokenClaims {
-	token := c.GetHeader("X-Subject-Token")
-	if token == "" {
-		authHeader := c.GetHeader("Authorization")
-		if len(authHeader) > 7 && authHeader[:7] == "Bearer " {
-			token = authHeader[7:]
-		}
-	}
-
-	if token == "" {
-		return nil
-	}
-
-	claims, err := h.jwtService.ValidateToken(token)
-	if err != nil {
-		return nil
-	}
-
-	return claims
 }
 
 // HandleList returns a list of distributions accessible to the current user
@@ -102,7 +33,7 @@ func (h *Handler) HandleList(c *gin.Context) {
 		statusFilter = &status
 	}
 
-	claims := h.getTokenClaims(c)
+	claims := common.GetTokenClaimsFromRequest(c, h.jwtService)
 	var userID string
 	var isAdmin bool
 	if claims != nil {
@@ -221,7 +152,7 @@ func (h *Handler) HandleGet(c *gin.Context) {
 		return
 	}
 
-	claims := h.getTokenClaims(c)
+	claims := common.GetTokenClaimsFromRequest(c, h.jwtService)
 	var userID string
 	var isAdmin bool
 	if claims != nil {
@@ -443,7 +374,7 @@ func (h *Handler) HandleGetLogs(c *gin.Context) {
 		return
 	}
 
-	claims := h.getTokenClaims(c)
+	claims := common.GetTokenClaimsFromRequest(c, h.jwtService)
 	var userID string
 	var isAdmin bool
 	if claims != nil {
