--- src/ldfd/auth/user.go
@@ -1,13 +1,0 @@
-package auth
-
-import "database/sql"
-
-// UserManager handles user, role, and token persistence
-type UserManager struct {
-	db *sql.DB
-}
-
-// NewUserManager creates a new user manager
-func NewUserManager(db *sql.DB) *UserManager {
-	return &UserManager{db: db}
-}
